+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/intellij/report/ConsoleReportBuilder.kt
@@ -21,6 +21,10 @@ internal class ConsoleReportBuilder(
                 coverageDataProvider = dataProvider
                 outputStream = System.out
                 shrinkLongClassName = true
+
+                targetInstr(0)
+                targetLines(0)
+                targetBranches(0)
             }
             TextualReportFacade.generateReport(buildContext)
         }