+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/textual/RawCoverageData.kt
@@ -1,51 +1,66 @@
 package io.github.surpsg.deltacoverage.report.textual
 
+import io.github.surpsg.deltacoverage.config.CoverageEntity
+import io.github.surpsg.deltacoverage.report.textual.Coverage.Companion.has
+
 internal class RawCoverageData private constructor(
     val aClass: String,
-    val branchesCovered: Int,
-    val branchesTotal: Int,
-    val linesCovered: Int,
-    val linesTotal: Int,
+
+    val instr: Coverage,
+    val lines: Coverage,
+    val branches: Coverage,
 ) {
 
-    val branchesRatio: Double
-        get() = branchesCovered ratioBy branchesTotal
-
-    val linesRatio: Double
-        get() = linesCovered ratioBy linesTotal
-
-    private infix fun Int.ratioBy(base: Int): Double = this.toDouble() / base
+    init {
+        require(instr.entity == CoverageEntity.INSTRUCTION)
+        require(lines.entity == CoverageEntity.LINE)
+        require(branches.entity == CoverageEntity.BRANCH)
+    }
 
     fun merge(other: RawCoverageData) = this.let { thisData ->
         RawCoverageData {
             aClass = thisData.aClass
 
-            branchesCovered = thisData.branchesCovered + other.branchesCovered
-            branchesTotal = thisData.branchesTotal + other.branchesTotal
-
-            linesCovered = thisData.linesCovered + other.linesCovered
-            linesTotal = thisData.linesTotal + other.linesTotal
+            thisData.instr.mergeWith(other.instr).let {
+                instr(it.covered, it.total)
+            }
+            thisData.lines.mergeWith(other.lines).let {
+                lines(it.covered, it.total)
+            }
+            thisData.branches.mergeWith(other.branches).let {
+                branches(it.covered, it.total)
+            }
         }
     }
 
     internal class Builder {
         lateinit var aClass: String
 
-        var branchesCovered: Int? = null
-        var branchesTotal: Int? = null
+        private var instr: Coverage? = null
+        private var lines: Coverage? = null
+        private var branches: Coverage? = null
 
-        var linesCovered: Int? = null
-        var linesTotal: Int? = null
+        fun instr(covered: Int, total: Int) {
+            instr = CoverageEntity.INSTRUCTION.has(covered, total)
+        }
 
-        fun build() = RawCoverageData(
-            aClass = aClass,
+        fun lines(covered: Int, total: Int) {
+            lines = CoverageEntity.LINE.has(covered, total)
+        }
 
-            branchesCovered = requireNotNull(branchesCovered),
-            branchesTotal = requireNotNull(branchesTotal),
+        fun branches(covered: Int, total: Int) {
+            branches = CoverageEntity.BRANCH.has(covered, total)
+        }
 
-            linesCovered = requireNotNull(linesCovered),
-            linesTotal = requireNotNull(linesTotal),
-        )
+        fun build(): RawCoverageData {
+            return RawCoverageData(
+                aClass = aClass,
+
+                instr = requireNotNull(instr),
+                branches = requireNotNull(branches),
+                lines = requireNotNull(lines),
+            )
+        }
     }
 
     companion object {
@@ -57,10 +72,10 @@ internal class RawCoverageData private constructor(
             customize: Builder.() -> Unit = {}
         ) = RawCoverageData {
             aClass = ""
-            branchesCovered = 0
-            branchesTotal = 0
-            linesCovered = 0
-            linesTotal = 0
+
+            instr(0, 0)
+            lines(0, 0)
+            branches(0, 0)
 
             customize()
         }