+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/textual/TextualReportFacade.kt
@@ -1,24 +1,27 @@
 package io.github.surpsg.deltacoverage.report.textual
 
+import io.github.surpsg.deltacoverage.config.CoverageEntity
 import io.github.surpsg.deltacoverage.lib.kotlin.applyIf
 import io.github.surpsg.deltacoverage.report.ReportBound
 import io.github.surpsg.deltacoverage.report.ReportType
+import io.github.surpsg.deltacoverage.report.textual.Coverage.Companion.toPercents
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.BRANCHES_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.CLASS_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.FAILURE_COV_PATTERN
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.HEADERS
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.INDEXED_HEADERS
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.INSTR_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.LINES_H
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.MAX_CLASS_COLUMN_LENGTH
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.NA_VALUE
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.PERCENT_MULTIPLIER
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.SHRINK_PLACEHOLDER
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.SUCCESS_COV_PATTERN
+import io.github.surpsg.deltacoverage.report.textual.ReportsConstants.TARGET_COV_PATTERN
 import io.github.surpsg.deltacoverage.report.textual.TextualReportRenderer.Context
 import java.io.OutputStream
 
 internal object TextualReportFacade {
-
-    private const val MAX_CLASS_COLUMN_LENGTH = 100
-    private const val SHRINK_PLACEHOLDER = "..."
-
-    private const val PERCENT_MULTIPLIER = 100
-
-    private const val NA_VALUE = ""
-
-    private const val CLASS_H = "Class"
-    private const val LINES_H = "Lines"
-    private const val BRANCHES_H = "Branches"
-    private val HEADERS = listOf(CLASS_H, LINES_H, BRANCHES_H)
 
     fun generateReport(
         buildContext: BuildContext,
@@ -26,7 +29,7 @@ internal object TextualReportFacade {
         val rawCoverageData: List<RawCoverageData> = buildContext.coverageDataProvider.obtainData()
         val coverageDataValues: List<List<String>> =
             rawCoverageData
-                .sortedByDescending { it.linesRatio }
+                .sortedByDescending { it.lines.ratio }
                 .asSequence()
                 .map { it.toValuesCollection(buildContext) }
                 .toList()
@@ -36,29 +39,63 @@ internal object TextualReportFacade {
                 output = buildContext.outputStream
                 title = buildTitle(buildContext)
                 headers = HEADERS
-                footer = rawCoverageData.computeTotal().toValuesCollection(buildContext)
+                footer = buildFooter(rawCoverageData, buildContext)
                 rows = coverageDataValues
             }
         )
     }
 
+    private fun buildFooter(
+        rawCoverageData: List<RawCoverageData>,
+        buildContext: BuildContext,
+    ): List<List<String>> {
+        val expectedValuesLine = Array(INDEXED_HEADERS.size) { "" }
+        sequenceOf(
+            BRANCHES_H to CoverageEntity.BRANCH,
+            LINES_H to CoverageEntity.LINE,
+            INSTR_H to CoverageEntity.INSTRUCTION,
+        )
+            .filter { (_, entity) -> buildContext.targetCoverage.getValue(entity) > 0 }
+            .forEach { (headerName, entity) ->
+                val expectedCoveragePercents: Int = buildContext.targetCoverage.getValue(entity)
+                val headerIndex = INDEXED_HEADERS.getValue(headerName)
+                expectedValuesLine[headerIndex] = TARGET_COV_PATTERN.format(expectedCoveragePercents)
+            }
+
+        return listOf(
+            rawCoverageData.computeTotal().toValuesCollection(buildContext) { coverage, formatted ->
+                val expectedCoveragePercents: Int = buildContext.targetCoverage.getValue(coverage.entity)
+                when {
+                    expectedCoveragePercents < 1 -> formatted
+                    (coverage.ratio.toPercents() < expectedCoveragePercents) -> FAILURE_COV_PATTERN.format(formatted)
+                    else -> SUCCESS_COV_PATTERN.format(formatted)
+                }
+            },
+            expectedValuesLine.toList(),
+        )
+    }
+
     private fun RawCoverageData.toValuesCollection(
         buildContext: BuildContext,
-    ): List<String> = HEADERS.map { header ->
+        decorateCoverageValue: (Coverage, String) -> String = { _, formatted -> formatted },
+    ): List<String> = INDEXED_HEADERS.keys.map { header ->
         when (header) {
-            LINES_H -> linesRatio.formatToPercentage()
-
             CLASS_H -> aClass.applyIf(buildContext.shrinkLongClassName) {
                 shrinkClassName(MAX_CLASS_COLUMN_LENGTH)
             }
 
+            LINES_H -> decorateCoverageValue(lines, lines.ratio.formatToPercentage())
+
             BRANCHES_H -> {
-                if (branchesTotal == 0) {
+                if (branches.total == 0) {
                     NA_VALUE
                 } else {
-                    branchesRatio.formatToPercentage()
+                    decorateCoverageValue(branches, branches.ratio.formatToPercentage())
                 }
             }
+
+            INSTR_H -> decorateCoverageValue(instr, instr.ratio.formatToPercentage())
+
 
             else -> error("Unknown header: $header")
         }
@@ -100,6 +137,7 @@ internal object TextualReportFacade {
         val coverageDataProvider: RawCoverageDataProvider,
         val outputStream: OutputStream,
         val shrinkLongClassName: Boolean,
+        val targetCoverage: Map<CoverageEntity, Int>,
     ) {
 
         init {
@@ -118,17 +156,49 @@ internal object TextualReportFacade {
             lateinit var outputStream: OutputStream
             var shrinkLongClassName: Boolean = false
 
-            fun build() = BuildContext(
-                reportType,
-                reportBound,
-                coverageDataProvider,
-                outputStream,
-                shrinkLongClassName
-            )
+            private var targetInstr: TargetCoverage? = null
+            private var targetLines: TargetCoverage? = null
+            private var targetBranches: TargetCoverage? = null
+
+            fun targetInstr(percents: Int) {
+                targetInstr = TargetCoverage(CoverageEntity.INSTRUCTION, percents)
+            }
+
+            fun targetLines(percents: Int) {
+                targetLines = TargetCoverage(CoverageEntity.LINE, percents)
+            }
+
+            fun targetBranches(percents: Int) {
+                targetBranches = TargetCoverage(CoverageEntity.BRANCH, percents)
+            }
+
+            fun build(): BuildContext {
+                return BuildContext(
+                    reportType,
+                    reportBound,
+                    coverageDataProvider,
+                    outputStream,
+                    shrinkLongClassName,
+                    sequenceOf(
+                        CoverageEntity.INSTRUCTION to targetInstr,
+                        CoverageEntity.LINE to targetLines,
+                        CoverageEntity.BRANCH to targetBranches
+                    ).map { (entity, targetCov) ->
+                        requireNotNull(targetCov) {
+                            "Required $entity was null."
+                        }
+                    }.associate { it.coverageEntity to it.percents }
+                )
+            }
         }
 
         companion object {
             operator fun invoke(block: Builder.() -> Unit): BuildContext = Builder().apply(block).build()
         }
     }
+
+    private data class TargetCoverage(
+        val coverageEntity: CoverageEntity,
+        val percents: Int,
+    )
 }