+++ delta-coverage-core/src/test/kotlin/io/github/surpsg/deltacoverage/report/intellij/coverage/IntellijRawCoverageDataProviderTest.kt
@@ -2,6 +2,8 @@ package io.github.surpsg.deltacoverage.report.intellij.coverage
 
 import com.intellij.rt.coverage.data.LineData
 import com.intellij.rt.coverage.data.ProjectData
+import com.intellij.rt.coverage.data.instructions.ClassInstructions
+import com.intellij.rt.coverage.data.instructions.LineInstructions
 import io.github.surpsg.deltacoverage.report.textual.RawCoverageData
 import io.kotest.assertions.assertSoftly
 import io.kotest.matchers.collections.shouldBeEmpty
@@ -31,11 +33,22 @@ class IntellijRawCoverageDataProviderTest {
         val sourceName = "Class1.java"
         val provider = IntellijRawCoverageDataProvider(
             ProjectData().apply {
+                instructions[className] = ClassInstructions(
+                    arrayOf(
+                        LineInstructions().apply {
+                            instructions = 3
+                        },
+                        LineInstructions().apply {
+                            instructions = 5
+                        }
+                    )
+                )
+
                 getOrCreateClassData(className).apply {
                     source = sourceName
 
                     setLines(
-                        arrayOf(LineData(1, ""))
+                        arrayOf(LineData(1, "").apply { hits = 1 })
                     )
                 }
             }
@@ -49,10 +62,9 @@ class IntellijRawCoverageDataProviderTest {
             size shouldBe 1
             first() shouldBeEqualToComparingFields RawCoverageData {
                 aClass = className
-                branchesTotal = 0
-                branchesCovered = 0
-                linesCovered = 0
-                linesTotal = 1
+                branches(0, 0)
+                lines(1, 1)
+                instr(5, 5)
             }
         }
     }