+++ delta-coverage-core/src/test/kotlin/io/github/surpsg/deltacoverage/report/textual/RawCoverageDataTest.kt
@@ -1,6 +1,5 @@
 package io.github.surpsg.deltacoverage.report.textual
 
-import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.equality.shouldBeEqualToComparingFields
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
@@ -23,12 +22,11 @@ class RawCoverageDataTest {
     ) {
         // GIVEN
         val data = RawCoverageData.newBlank {
-            branchesCovered = covered
-            branchesTotal = total
+            branches(covered = covered, total = total)
         }
 
         // WHEN // THEN
-        data.branchesRatio shouldBe expected
+        data.branches.ratio shouldBe expected
     }
 
     @ParameterizedTest
@@ -44,12 +42,31 @@ class RawCoverageDataTest {
     ) {
         // GIVEN
         val data = RawCoverageData.newBlank {
-            branchesCovered = covered
-            branchesTotal = total
+            lines(covered = covered, total = total)
         }
 
         // WHEN // THEN
-        data.branchesRatio shouldBe expected
+        data.lines.ratio shouldBe expected
+    }
+
+    @ParameterizedTest
+    @CsvSource(
+        "20, 100, 0.2",
+        "4, 16, 0.25",
+        "10, 10, 1.0",
+    )
+    fun `should return correct ration for instructions`(
+        covered: Int,
+        total: Int,
+        expected: Double,
+    ) {
+        // GIVEN
+        val data = RawCoverageData.newBlank {
+            instr(covered = covered, total = total)
+        }
+
+        // WHEN // THEN
+        data.instr.ratio shouldBe expected
     }
 
     @Test
@@ -57,18 +74,16 @@ class RawCoverageDataTest {
         // GIVEN
         val data1 = RawCoverageData {
             aClass = "class1"
-            branchesCovered = 5
-            branchesTotal = 10
-            linesCovered = 5
-            linesTotal = 10
+            instr(4, 4)
+            branches(5, 10)
+            lines(5, 10)
         }
 
         val data2 = RawCoverageData {
             aClass = "class2"
-            branchesCovered = 5
-            branchesTotal = 10
-            linesCovered = 5
-            linesTotal = 10
+            instr(2, 6)
+            branches(5, 10)
+            lines(5, 10)
         }
 
         // WHEN
@@ -77,38 +92,9 @@ class RawCoverageDataTest {
         // THEN
         mergedData shouldBeEqualToComparingFields RawCoverageData {
             aClass = "class1"
-            branchesCovered = 10
-            branchesTotal = 20
-            linesCovered = 10
-            linesTotal = 20
-        }
-    }
-
-    @ParameterizedTest
-    @CsvSource(
-        nullValues = ["null"],
-        value = [
-            "null, 1, 1, 1",
-            "1, null, 1, 1",
-            "1, 1, null, 1",
-            "1, 1, 1, null",
-        ]
-    )
-    fun `should throw if fields not set`(
-        branchesCovered: Int?,
-        branchesTotal: Int?,
-        linesCovered: Int?,
-        linesTotal: Int?,
-    ) {
-
-        shouldThrow<IllegalArgumentException> {
-            RawCoverageData {
-                aClass = "class"
-                this.branchesCovered = branchesCovered
-                this.branchesTotal = branchesTotal
-                this.linesCovered = linesCovered
-                this.linesTotal = linesTotal
-            }
+            instr(6, 10)
+            branches(10, 20)
+            lines(10, 20)
         }
     }
 }